app/core/security.py
from datetime import datetime, timedelta

from jose import jwt

SECRET_KEY = "CHANGE_ME"
ALGORITHM = "HS256"


def create_access_token(user_id: str):
    payload = {
        "sub": user_id,
        "exp": datetime.utcnow() + timedelta(hours=24),
    }
    return jwt.encode(payload, SECRET_KEY, algorithm=ALGORITHM)
