requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-jose
passlib[bcrypt]
python-dotenv
